requirements.txt
nbtlib~=2.0.4
attrs~=21.4.0
PyYAML~=6.0